Create Falling Star | JavaScript
My goal in this project was to make some type of artwork based on a song by one of my favorite musical artists. This work of art was based on the song "Falling Star" by a rapper who goes by the name of "NoCap". This was the 2nd coding project that I ever completed.