Create Spaceman Game | JavaScript
This was a video game that I created during my time at CodeHouse. It was made after less than a month of coding experience and it's my personal favorite project so far. The game isn't perfect and has a few bugs, but I'm more proud of myself for this project than I am for any of my other personal creations.

Game Instructions:
You're a spaceman who now finds thousands of asteroids falling towards him! Use the arrow keys in order to move the spaceman up and down or side to side to avoid the asteroids. Avoid 30 asteroids in order to win, but keep going to get a high as a score that you can if possible! Whenever you die, refresh the page to restart the game.